--- DfE.GIAP.All/DfE.GIAP.Core.UnitTests/Content/CompositionRootTests.cs
@@ -1,44 +1,0 @@
-﻿using DfE.GIAP.Core.Common.Application;
-using DfE.GIAP.Core.Common.CrossCutting;
-using DfE.GIAP.Core.Content.Application.Options;
-using DfE.GIAP.Core.Content.Application.Repository;
-using DfE.GIAP.Core.Content.Application.UseCases.GetContentByPageKeyUseCase;
-using DfE.GIAP.Core.Content.Infrastructure.Repositories;
-using DfE.GIAP.Core.SharedTests;
-using DfE.GIAP.Core.SharedTests.TestDoubles;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
-using CompositionRoot = DfE.GIAP.Core.Content.CompositionRoot;
-
-namespace DfE.GIAP.Core.UnitTests.Content;
-public sealed class CompositionRootTests
-{
-    [Fact]
-    public void ThrowsArgumentNullException_When_ServicesIsNull()
-    {
-        IServiceCollection? serviceCollection = null;
-        ConfigurationTestDoubles.Default();
-        Action register = () => CompositionRoot.AddContentDependencies(serviceCollection);
-        Assert.Throws<ArgumentNullException>(register);
-    }
-
-    [Fact]
-    public void Registers_CompositionRoot_CanResolve_Services()
-    {
-        // Arrange
-        IServiceCollection services = ServiceCollectionTestDoubles.Default().AddSharedDependencies();
-
-        // Act
-        IServiceCollection registeredServices = CompositionRoot.AddContentDependencies(services);
-        IServiceProvider provider = registeredServices.BuildServiceProvider();
-
-        // Assert
-        Assert.NotNull(registeredServices);
-        Assert.NotNull(provider);
-
-        Assert.NotNull(provider.GetService<IUseCase<GetContentByPageKeyUseCaseRequest, GetContentByPageKeyUseCaseResponse>>());
-        Assert.NotNull(provider.GetService<IMapper<ContentDto, Core.Content.Application.Model.Content>>());
-        Assert.NotNull(provider.GetService<IContentReadOnlyRepository>());
-        Assert.NotNull(provider.GetService<IOptions<PageContentOptions>>());
-    }
-}
