--- DfE.GIAP.All/DfE.GIAP.Core.UnitTests/Content/TestDoubles/ContentReadOnlyRepositoryTestDoubles.cs
@@ -1,7 +1,0 @@
-﻿using DfE.GIAP.Core.Content.Application.Repository;
-
-namespace DfE.GIAP.Core.UnitTests.Content.TestDoubles;
-internal sealed class ContentReadOnlyRepositoryTestDoubles
-{
-    internal static Mock<IContentReadOnlyRepository> Default() => new();
-}
