--- DfE.GIAP.All/DfE.GIAP.Core.UnitTests/Content/TestDoubles/PageContentOptionTestDoubles.cs
@@ -1,13 +1,0 @@
-﻿using DfE.GIAP.Core.Content.Application.Options;
-
-namespace DfE.GIAP.Core.UnitTests.Content.TestDoubles;
-internal static class PageContentOptionTestDoubles
-{
-    public static PageContentOption Default() => new();
-    public static PageContentOption StubFor(string documentId)
-    {
-        PageContentOption stub = Default();
-        stub.DocumentId = documentId;
-        return stub;
-    }
-}
