--- DfE.GIAP.All/DfE.GIAP.Core.UnitTests/Content/TestDoubles/PageContentOptionsProviderTestDoubles.cs
@@ -1,7 +1,0 @@
-﻿using DfE.GIAP.Core.Content.Application.Options.Provider;
-
-namespace DfE.GIAP.Core.UnitTests.Content.TestDoubles;
-internal static class PageContentOptionsProviderTestDoubles
-{
-    internal static Mock<IPageContentOptionsProvider> Default() => new();
-}
