+++ DfE.GIAP.All/DfE.GIAP.Core.UnitTests/Contents/TestDoubles/PageContentOptionsProviderTestDoubles.cs
@@ -1,0 +1,7 @@
+﻿using DfE.GIAP.Core.Contents.Application.Options.Provider;
+
+namespace DfE.GIAP.Core.UnitTests.Contents.TestDoubles;
+internal static class PageContentOptionsProviderTestDoubles
+{
+    internal static Mock<IPageContentOptionsProvider> Default() => new();
+}
