+++ DfE.GIAP.All/DfE.GIAP.Core.UnitTests/NewsArticles/UseCases/GetNewsArticles/GetNewsArticlesUseCaseTests.cs
@@ -1,6 +1,5 @@
 using DfE.GIAP.Core.NewsArticles.Application.Enums;
 using DfE.GIAP.Core.NewsArticles.Application.UseCases.GetNewsArticles;
-using DfE.GIAP.Core.UnitTests.NewsArticles.UseCases.GetNewsArticlesById.TestDoubles;
 
 namespace DfE.GIAP.Core.UnitTests.NewsArticles.UseCases.GetNewsArticles;
