--- DfE.GIAP.All/DfE.GIAP.Core/Content/Application/Options/PageContentOption.cs
@@ -1,12 +1,0 @@
-﻿namespace DfE.GIAP.Core.Content.Application.Options;
-
-/// <summary>
-/// Represents configuration options for a specific page, including the associated document ID.
-/// </summary>
-public sealed class PageContentOption
-{
-    /// <summary>
-    /// Gets or sets the ID of the document associated with the page.
-    /// </summary>
-    public string DocumentId { get; set; }
-}
