--- DfE.GIAP.All/DfE.GIAP.Core/Content/Application/Options/PageContentOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace DfE.GIAP.Core.Content.Application.Options;
-
-/// <summary>
-/// A dictionary mapping page keys to their corresponding <see cref="PageContentOption"/>.
-/// Used to configure content options for multiple pages.
-/// </summary>
-public class PageContentOptions : Dictionary<string, PageContentOption>
-{
-}
