+++ DfE.GIAP.All/DfE.GIAP.Core/Contents/Application/Options/PageContentOption.cs
@@ -1,0 +1,12 @@
+﻿namespace DfE.GIAP.Core.Contents.Application.Options;
+
+/// <summary>
+/// Represents configuration options for a specific page, including the associated document ID.
+/// </summary>
+public sealed class PageContentOption
+{
+    /// <summary>
+    /// Gets or sets the ID of the document associated with the page.
+    /// </summary>
+    public string DocumentId { get; set; }
+}
