+++ DfE.GIAP.All/DfE.GIAP.Core/Contents/Application/Options/PageContentOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace DfE.GIAP.Core.Contents.Application.Options;
+
+/// <summary>
+/// A dictionary mapping page keys to their corresponding <see cref="PageContentOption"/>.
+/// Used to configure content options for multiple pages.
+/// </summary>
+public class PageContentOptions : Dictionary<string, PageContentOption>
+{
+}
