+++ DfE.GIAP.All/DfE.GIAP.Core/Contents/Application/Repository/IContentReadOnlyRepository.cs
@@ -1,0 +1,17 @@
+﻿using DfE.GIAP.Core.Contents.Application.Models;
+
+namespace DfE.GIAP.Core.Contents.Application.Repository;
+
+/// <summary>
+/// Defines a read-only repository interface for retrieving content by ID.
+/// </summary>
+public interface IContentReadOnlyRepository
+{
+    /// <summary>
+    /// Asynchronously retrieves content by its unique identifier.
+    /// </summary>
+    /// <param name="id">The unique identifier of the content.</param>
+    /// <param name="ctx">Optional cancellation token.</param>
+    /// <returns>The content associated with the specified ID.</returns>
+    Task<Content> GetContentByIdAsync(string id, CancellationToken ctx = default);
+}
