+++ DfE.GIAP.All/DfE.GIAP.Web/ViewModels/LandingViewModel.cs
@@ -1,4 +1,4 @@
-﻿using DfE.GIAP.Core.Content.Application.Models;
+﻿using DfE.GIAP.Core.Contents.Application.Models;
 using System.Diagnostics.CodeAnalysis;
 namespace DfE.GIAP.Web.ViewModels;
